# Conflicts:
#	registro-cantidad-matriculas.html

+++ assets/components/PHP_Consultas/Registro_Area_Academica_Licenciatura/Eliminar_Registro.php
@@ -1,15 +1,28 @@
 <?php
 
-require_once "../Conexion.php";
+require "../Conexion.php";
+require_once "../Usuarios/Verificar_Permisos_Usuarios.php";
 $conexion = conexion();
+$conn = conexion();
+session_start();
+$id_usuario = $_SESSION["id_usuario"];
+$tabla = 'cuerpos_academicos';
+$stmt = consultaPermisos($conn,$id_usuario,$tabla,'Delete');
 
-$id_cuerpo_academico=$_POST['id_cuerpo_academico'];
+$stmt->execute();
 
-$stmt = $conexion->prepare("delete from cuerpos_academicos where ID_CUERPO_ACADEMICO=?");
-$stmt->bind_param("i",$id_cuerpo_academico);
+if($stmt->fetch()) {
+    $id_cuerpo_academico = $_POST['id_cuerpo_academico'];
 
-echo $resultado = $stmt->execute();
-$stmt->close();
-$conexion->close();
+    $stmt = $conexion->prepare("delete from cuerpos_academicos where ID_CUERPO_ACADEMICO=?");
+    $stmt->bind_param("i", $id_cuerpo_academico);
+
+    echo $resultado = $stmt->execute();
+    $stmt->close();
+    $conexion->close();
+}else{
+        echo 2;
+    }
+
 
 ?>
